requirements.txt
Flask==2.2.2
numpy>=1.9.2
pandas>=0.19
werkzeug==2.2.2
tensorflow-cpu==2.9.*
tensorflow-text==2.9.*
stop-words==2018.7.23 
joblib==1.2.0
scikit-learn>=1.0.2
nltk==3.7

